feat: footer 구현
