+++ src/components/ContactList/contactList.module.scss
@@ -3,10 +3,10 @@
   align-items: center;
 
   li + li {
-    margin-left: 20px;
+    margin-left: 12px;
   }
 
   svg {
-    fill: var(--color-text);
+    fill: var(--color-svg);
   }
 }
