+++ src/routes/Portfolio/Aside/aside.module.scss
@@ -8,16 +8,15 @@
     position: relative;
     display: block;
     height: 100%;
-    padding: 0 20px;
+    padding: 20px;
     color: var(--color-text);
     background: var(--primary);
     transition: background 0.2s ease-in 0s;
 
     h1 {
-      height: 80px;
+      height: 60px;
       font-size: 24px;
       font-weight: 600;
-      line-height: 80px;
     }
 
     .intro {
@@ -25,17 +24,12 @@
       line-height: 1.6;
     }
 
-    .contact {
-      display: flex;
+    & > ul {
       margin-top: 40px;
+    }
 
-      li + li {
-        margin-left: 20px;
-      }
-
-      svg {
-        fill: var(--color-text);
-      }
+    .copyright {
+      margin-top: 20px;
     }
 
     .buttonsAside {
